refactor: extract TableFilter from oss to core
